# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Sampooran Holidays Travel Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Country {
  id: number;
  name: string;
  slug: string;
  code: string;
  imageUrl?: string;
  packageCount: number;
}

export interface State {
  id: number;
  name: string;
  slug: string;
  countryId: number;
  countryName: string;
  imageUrl?: string;
  packageCount: number;
  description?: string;
}

export interface Destination {
  id: number;
  name: string;
  slug: string;
  stateId: number;
  stateName: string;
  countryName: string;
  imageUrl?: string;
  thumbnailUrl?: string;
  description?: string;
  isFeatured: boolean;
  packageCount: number;
  bestTimeToVisit?: string;
  altitude?: string;
  temperature?: string;
}

export type PackagePackageType =
  (typeof PackagePackageType)[keyof typeof PackagePackageType];

export const PackagePackageType = {
  b2b: "b2b",
  b2c: "b2c",
  both: "both",
} as const;

export interface Package {
  id: number;
  name: string;
  slug: string;
  destinationId: number;
  destinationName: string;
  stateName: string;
  countryName: string;
  imageUrl?: string;
  thumbnailUrl?: string;
  shortDescription?: string;
  duration: number;
  nights: number;
  pricePerPerson: number;
  originalPrice?: number;
  discountPercent?: number;
  category: string;
  packageType: PackagePackageType;
  isFeatured: boolean;
  rating?: number;
  reviewCount?: number;
  highlights?: string[];
  cities?: string[];
  tags?: string[];
  inclusionIcons?: string[];
}

export type DestinationDetail = Destination & {
  longDescription?: string;
  highlights?: string[];
  thingsToDo?: string[];
  howToReach?: string;
  localAttractions?: string[];
  localCuisine?: string[];
  accommodation?: string[];
  travelTips?: string[];
  famousFor?: string[];
  activities?: string[];
  festivals?: string[];
  safetyInfo?: string;
  faqs?: Faq[];
  galleryImages?: string[];
  nearestAirport?: string;
  nearestRailway?: string;
  distanceFromCapital?: string;
  packages: Package[];
};

export interface ItineraryDay {
  day: number;
  title: string;
  description: string;
  activities?: string[];
  meals?: string[];
  accommodation?: string;
}

export interface HotelOption {
  city: string;
  hotelName: string;
  category: string;
  nights: number;
  roomType?: string;
}

export interface Faq {
  question: string;
  answer: string;
}

export type PackageDetail = Package & {
  longDescription?: string;
  itinerary: ItineraryDay[];
  inclusions: string[];
  exclusions: string[];
  hotels?: HotelOption[];
  tourType?: string;
  groupSize?: string;
  pickupPoint?: string;
  importantNotes?: string[];
  cancellationPolicy?: string;
  faqs?: Faq[];
  relatedPackages?: Package[];
};

export interface PackageStats {
  totalPackages: number;
  totalDestinations: number;
  totalStates: number;
  totalCountries: number;
  happyTravelers: number;
  yearsExperience: number;
  avgRating: number;
}

export type TransportServiceType =
  (typeof TransportServiceType)[keyof typeof TransportServiceType];

export const TransportServiceType = {
  taxi: "taxi",
  tempo: "tempo",
  bus: "bus",
} as const;

export interface TransportService {
  id: number;
  name: string;
  slug: string;
  type: TransportServiceType;
  vehicleModel?: string;
  capacity: number;
  imageUrl?: string;
  pricePerKm?: number;
  basePrice?: number;
  description?: string;
  features?: string[];
  isAC?: boolean;
}

export interface TransportRoute {
  id: number;
  from: string;
  to: string;
  distance: number;
  estimatedTime: string;
  startingPrice: number;
  isPopular: boolean;
}

export type InquiryInputInquiryType =
  (typeof InquiryInputInquiryType)[keyof typeof InquiryInputInquiryType];

export const InquiryInputInquiryType = {
  package: "package",
  transport: "transport",
  customized: "customized",
  b2b: "b2b",
} as const;

export interface InquiryInput {
  name: string;
  email?: string;
  phone: string;
  inquiryType: InquiryInputInquiryType;
  packageId?: number;
  destination?: string;
  travelDate?: string;
  numberOfPersons?: number;
  message: string;
  budget?: number;
}

export interface Inquiry {
  id: number;
  name: string;
  email?: string;
  phone: string;
  inquiryType: string;
  packageId?: number;
  destination?: string;
  travelDate?: string;
  numberOfPersons?: number;
  message: string;
  budget?: number;
  status: string;
  createdAt: string;
}

export interface Testimonial {
  id: number;
  name: string;
  location?: string;
  avatarUrl?: string;
  rating: number;
  review: string;
  packageName?: string;
  travelDate?: string;
  isFeatured: boolean;
}

export interface BlogPost {
  id: number;
  title: string;
  slug: string;
  excerpt: string;
  content?: string;
  imageUrl?: string;
  author: string;
  publishedAt: string;
  tags?: string[];
  readTime?: number;
}

export type ListDestinationsParams = {
  country?: string;
  state?: string;
  featured?: boolean;
  limit?: number;
  offset?: number;
};

export type ListDestinations200 = {
  destinations: Destination[];
  total: number;
};

export type GetFeaturedDestinations200 = {
  destinations: Destination[];
};

export type ListCountries200 = {
  countries: Country[];
};

export type ListStatesParams = {
  country?: string;
};

export type ListStates200 = {
  states: State[];
};

export type ListPackagesParams = {
  destinationSlug?: string;
  state?: string;
  country?: string;
  category?: string;
  minDays?: number;
  maxDays?: number;
  minPrice?: number;
  maxPrice?: number;
  type?: ListPackagesType;
  featured?: boolean;
  limit?: number;
  offset?: number;
  search?: string;
};

export type ListPackagesType =
  (typeof ListPackagesType)[keyof typeof ListPackagesType];

export const ListPackagesType = {
  b2b: "b2b",
  b2c: "b2c",
  both: "both",
} as const;

export type ListPackages200 = {
  packages: Package[];
  total: number;
};

export type GetFeaturedPackages200 = {
  packages: Package[];
};

export type GetTrendingPackages200 = {
  packages: Package[];
};

export type ListTransportServicesParams = {
  type?: ListTransportServicesType;
  route?: string;
};

export type ListTransportServicesType =
  (typeof ListTransportServicesType)[keyof typeof ListTransportServicesType];

export const ListTransportServicesType = {
  taxi: "taxi",
  tempo: "tempo",
  bus: "bus",
} as const;

export type ListTransportServices200 = {
  services: TransportService[];
};

export type ListPopularRoutes200 = {
  routes: TransportRoute[];
};

export type ListTestimonialsParams = {
  featured?: boolean;
  limit?: number;
};

export type ListTestimonials200 = {
  testimonials: Testimonial[];
};

export type ListBlogPostsParams = {
  tag?: string;
  limit?: number;
  offset?: number;
};

export type ListBlogPosts200 = {
  posts: BlogPost[];
  total: number;
};