# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Sampooran Holidays Travel Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /destinations:
    get:
      operationId: listDestinations
      tags: [destinations]
      summary: List all destinations
      parameters:
        - name: country
          in: query
          schema:
            type: string
        - name: state
          in: query
          schema:
            type: string
        - name: featured
          in: query
          schema:
            type: boolean
        - name: limit
          in: query
          schema:
            type: integer
        - name: offset
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of destinations
          content:
            application/json:
              schema:
                type: object
                properties:
                  destinations:
                    type: array
                    items:
                      $ref: "#/components/schemas/Destination"
                  total:
                    type: integer
                required:
                  - destinations
                  - total
...
  /destinations/states:
    get:
      operationId: listStates
      tags: [destinations]
      summary: List states, optionally filtered by country
      parameters:
        - name: country
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of states
          content:
            application/json:
              schema:
                type: object
                properties:
                  states:
                    type: array
                    items:
                      $ref: "#/components/schemas/State"
                required:
                  - states
  /destinations/{slug}:
    get:
      operationId: getDestination
      tags: [destinations]
      summary: Get destination by slug
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Destination details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DestinationDetail"
        "404":
          description: Not found
...
  /blog:
    get:
      operationId: listBlogPosts
      tags: [blog]
      summary: List blog posts / travel tips
      parameters:
        - name: tag
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
        - name: offset
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Blog posts
          content:
            application/json:
              schema:
                type: object
                properties:
                  posts:
                    type: array
                    items:
                      $ref: "#/components/schemas/BlogPost"
                  total:
                    type: integer
                required:
                  - posts
                  - total
  /blog/{slug}:
    get:
      operationId: getBlogPost
      tags: [blog]
      summary: Get blog post by slug
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Blog post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlogPost"
        "404":
          description: Not found
components:
  schemas:
...
    State:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        countryId:
          type: integer
        countryName:
          type: string
        imageUrl:
          type: string
        packageCount:
          type: integer
        description:
          type: string
      required:
        - id
        - name
        - slug
        - countryId
        - countryName
        - packageCount
    Destination:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        stateId:
          type: integer
        stateName:
          type: string
        countryName:
          type: string
        imageUrl:
          type: string
        thumbnailUrl:
          type: string
        description:
          type: string
        isFeatured:
          type: boolean
        packageCount:
          type: integer
        bestTimeToVisit:
          type: string
        altitude:
          type: string
        temperature:
          type: string
      required:
        - id
        - name
        - slug
        - stateId
        - stateName
        - countryName
        - isFeatured
        - packageCount
    DestinationDetail:
      allOf:
        - $ref: "#/components/schemas/Destination"
        - type: object
          properties:
            longDescription:
              type: string
            highlights:
              type: array
              items:
                type: string
            thingsToDo:
              type: array
              items:
                type: string
            howToReach:
              type: string
            localAttractions:
              type: array
              items:
                type: string
            packages:
              type: array
              items:
                $ref: "#/components/schemas/Package"
          required:
            - packages
    Package:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        destinationId:
          type: integer
        destinationName:
          type: string
        stateName:
          type: string
        countryName:
          type: string
        imageUrl:
          type: string
        thumbnailUrl:
          type: string
        shortDescription:
          type: string
        duration:
          type: integer
        nights:
          type: integer
        pricePerPerson:
          type: number
        originalPrice:
          type: number
        discountPercent:
          type: integer
        category:
          type: string
        packageType:
          type: string
          enum: [b2b, b2c, both]
        isFeatured:
          type: boolean
        rating:
          type: number
        reviewCount:
          type: integer
        highlights:
          type: array
          items:
            type: string
        cities:
          type: array
          items:
            type: string
        tags:
          type: array
          items:
            type: string
        inclusionIcons:
          type: array
          items:
            type: string
      required:
        - id
        - name
        - slug
        - destinationId
        - destinationName
        - stateName
        - countryName
        - duration
        - nights
        - pricePerPerson
        - category
        - packageType
        - isFeatured
...
    BlogPost:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        slug:
          type: string
        excerpt:
          type: string
        content:
          type: string
        imageUrl:
          type: string
        author:
          type: string
        publishedAt:
          type: string
        tags:
          type: array
          items:
            type: string
        readTime:
          type: integer
      required:
        - id
        - title
        - slug
        - excerpt
        - author
        - publishedAt